Add files via upload
Kaggle Dataset: Whats Cooking
Some of our strongest geographic and cultural associations are tied to a region's local foods. This playground competitions asks you to predict the category of a dish's cuisine given a list of its ingredients. 
Submissions are evaluated on the categorization accuracy (the percent of dishes that you correctly classify).Submission should predict the cuisine for each recipe in the test set. 
Data description: 
In the dataset, recipe id, the type of cuisine, and the list of ingredients of each recipe (of variable length) is included. The data is stored in JSON format. In the test file test.json, the format of a recipe is the same as train.json, only the cuisine type is removed, asit is the target variable to be predicted.
Data Fields: 
train.json - the training set containing recipes id, type of cuisine, and list of ingredients
test.json - the test set containing recipes id, and list of ingredients
sample_submission.csv - a sample submission file in the correct format
Steps in Data Processing Conducted:
1. Removing space within words from the ingredients column 
2. Removing the list of lists within the 'ingredients' column
3. Creating an addditional column to remove the list of lists, hypen and then concatinating to orginal dataframe - now the new column is a cleaned column for applying TFIDF vectorization 
4.  Data was split into 80:20 for training and validation 
Models Applied: 
Support Vector Machine, 
Random Forest
Stochastic Gradient Descent
Error Metrics for evaluation – Accuracy

Model Results: 
LinearSVC performs better results than other algorithms used with 78 percent accuracy on test data.